requirements_gr.txt
neo4j
langchain
langchain_community
langchain_openai
langchain_groq
python-dotenv
ipykernel
langchain_neo4j
streamlit
langchain_experimental
lamini
agno
openai
google-generativeai
langchain-google-genai
duckduckgo-search
pypdf
lancedb
pandas
tantivy
yfinance
groq
mcp
google-genai
pydantic